Rename testing file.
