+++ packages/core/src/test.ts
@@ -1,5 +1,5 @@
-import { getRenderedStyles, purgeStyles } from '@aesthetic/style/lib/testing';
-import { design, lightTheme, darkTheme } from '@aesthetic/system/lib/testing';
+import { getRenderedStyles, purgeStyles } from '@aesthetic/style/lib/test';
+import { design, lightTheme, darkTheme } from '@aesthetic/system/lib/test';
 // eslint-disable-next-line unicorn/import-index
 import { registerDefaultTheme, resetForTesting } from './index';
