+++ packages/style/test.js
@@ -1,0 +1,1 @@
+module.exports = require('./lib/test.js');
