--- packages/system/testing.d.ts
@@ -1,1 +1,0 @@
-export * from './lib/testing';
